ref: move TraceablePeerConnection to RTC
also fixes lint errors
