+++ modules/xmpp/JingleSessionPC.js
@@ -3,7 +3,7 @@
 import {getLogger} from "jitsi-meet-logger";
 const logger = getLogger(__filename);
 var JingleSession = require("./JingleSession");
-var TraceablePeerConnection = require("./TraceablePeerConnection");
+var TraceablePeerConnection = require("../RTC/TraceablePeerConnection");
 var SDPDiffer = require("./SDPDiffer");
 var SDPUtil = require("./SDPUtil");
 var SDP = require("./SDP");
